% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCronjobs.R
\name{listCronjobs}
\alias{listCronjobs}
\title{List Loaded Cronjobs}
\usage{
listCronjobs()
}
\description{
Read your personal crontab and return a data.frame of information about jobs
added by \code{crontabR}.
}

